Refactor: di 폴더 생성
